requirements.txt
Django==1.6.4
Fabric==1.8.3
South==0.8.4
amqp==1.4.5
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.17
boto==2.28.0
celery==3.1.11
dj-database-url==0.3.0
dj-static==0.0.5
django-celery==3.1.10
django-cors-headers==0.12
django-storages==1.1.8
djangocms-admin-style==0.2.2
djangorestframework==2.3.13
ecdsa==0.11
gnureadline==6.3.3
gunicorn==18.0
ipython==2.0.0
kombu==3.0.16
matplotlib==1.3.1
nose==1.3.3
numpy==1.8.1
paramiko==1.12.4
psycopg2==2.5.2
pycrypto==2.6.1
pyparsing==2.0.2
pystache==0.5.4
python-dateutil==2.2
python-postmark==0.4.1
pytz==2014.2
readline==6.2.4.1
redis==2.9.1
six==1.6.1
static==1.0.2
tornado==3.2.1
wsgiref==0.1.2
pylibmc==1.2.3
django-pylibmc==0.5
pystmark==0.4.4
jsonfield==0.9.20
django-jsonfield==0.9.12
Markdown==2.4
django-filter==0.7
supervisor==3.0
